scanner: Move more code to the base class for better reuse
