+++ scanner/src/main/kotlin/scanners/ScanCode.kt
@@ -22,23 +22,18 @@ package com.here.ort.scanner.scanners
 import ch.frankel.slf4k.*
 import ch.qos.logback.classic.Level
 
-import com.here.ort.downloader.DownloadException
-import com.here.ort.downloader.Main
-import com.here.ort.model.Package
 import com.here.ort.scanner.ScanException
-import com.here.ort.scanner.ScanResultsCache
 import com.here.ort.scanner.Scanner
+import com.here.ort.scanner.ScannerResult
 import com.here.ort.util.ProcessCapture
 import com.here.ort.util.getCommandVersion
 import com.here.ort.util.jsonMapper
 import com.here.ort.util.log
-import com.here.ort.util.safeMkdirs
 
 import java.io.File
 
 object ScanCode : Scanner() {
 
-    private const val OUTPUT_EXTENSION = "json"
     private const val OUTPUT_FORMAT = "json-pp"
     private const val PROCESSES = 6
     private const val TIMEOUT = 300
@@ -46,40 +41,9 @@ object ScanCode : Scanner() {
     private val DEFAULT_OPTIONS = listOf("--copyright", "--license", "--info", "--diag", "--only-findings",
             "--strip-root")
 
-    override fun scan(pkg: Package, outputDirectory: File): Set<String> {
-        val scanResultsDirectory = File(outputDirectory, "scanResults").apply { safeMkdirs() }
-        val scannerName = javaClass.simpleName.toLowerCase()
-        val resultsFile = File(scanResultsDirectory,
-                "${pkg.name}-${pkg.version}_$scannerName.$OUTPUT_EXTENSION")
+    override val resultFileExtension = "json"
 
-        if (ScanResultsCache.read(pkg, resultsFile)) {
-            return parseLicenses(resultsFile)
-        }
-
-        val sourceDirectory = try {
-            val downloadDirectory = File(outputDirectory, "download").apply { safeMkdirs() }
-            Main.download(pkg, downloadDirectory)
-        } catch (e: DownloadException) {
-            if (Main.stacktrace) {
-                e.printStackTrace()
-            }
-
-            throw ScanException("Package '${pkg.identifier}' could not be scanned.", e)
-        }
-
-        return scanPath(sourceDirectory, resultsFile).also { ScanResultsCache.write(pkg, resultsFile) }
-    }
-
-    override fun scan(path: File, outputDirectory: File): Set<String> {
-        val scanResultsDirectory = File(outputDirectory, "scanResults").apply { safeMkdirs() }
-        val scannerName = javaClass.simpleName.toLowerCase()
-        val resultsFile = File(scanResultsDirectory,
-                "${path.nameWithoutExtension}_$scannerName.$OUTPUT_EXTENSION")
-
-        return scanPath(path, resultsFile)
-    }
-
-    private fun scanPath(path: File, resultsFile: File): Set<String> {
+    override fun scanPath(path: File, resultsFile: File): ScannerResult {
         val scancodeVersion = getCommandVersion("scancode", transform = {
             // "scancode --version" returns a string like "ScanCode version 2.0.1.post1.fb67a181", so remove the prefix.
             it.substringAfter("ScanCode version ")
@@ -105,7 +69,7 @@ object ScanCode : Scanner() {
         with(process) {
             if (exitValue() == 0) {
                 println("Stored ScanCode results in '${resultsFile.absolutePath}'.")
-                return parseLicenses(resultsFile)
+                return toScannerResult(resultsFile)
             } else {
                 throw ScanException(failMessage)
             }
@@ -116,7 +80,7 @@ object ScanCode : Scanner() {
         // TODO: Add results of license scan to YAML model
     }
 
-    private fun parseLicenses(resultsFile: File): Set<String> {
+    override fun toScannerResult(resultsFile: File): ScannerResult {
         val result = mutableSetOf<String>()
         val json = jsonMapper.readTree(resultsFile)
         val files = json["files"]
@@ -132,10 +96,6 @@ object ScanCode : Scanner() {
             }
         }
         return result
-    }
-
-    override fun toString(): String {
-        return javaClass.simpleName
     }
 
 }
